Emitter: Simpler printing of a call name on error.
I doubt anybody cares if something is a function or a variant. So that
part is gone now. The separator calculation is simpler too. It also
no longer writes directly into the msgbuf, and instead writes to
incoming strings. This, in turn, makes error printing easier.

Some tests were updated, since the name printing was altered as well
as the ordering of the name and the argument number.