Simplify calls with new tree_method used by parser.
Previously, call handling in emitter would need to check if the thing
being called was a method or not. This doesn't seem right, because it
then becomes more difficult to get a true argument count. As a result,
call handling became even more weird.

Now, parser will push things that are methods as tree_method, making
the calculation a lot simpler. This is a step toward making emitter's
call states simpler overall. No outward changes.