Emitter: Count arguments before pushing the first.
Now that methods are marked with tree_method, take advantage of to
easily determine if the first tree really provides a value. The tree's
value can be pushed later on, after checking if the argument count is
correct.

This does not do any external changes, but internally makes call
states less important.