Emitter: A better error message for 'Some' (no args passed).
Previously, the error message wouldn't check varargs (or optargs, if
that was even allowed in the future). So it was pretty lame.

I've also made a decision that if 'Some' is defined as '(A...)' that
invoking Some alone is not acceptable. This is to be consistent with
normal functions which need explicit invocation. Since variants are
not true functions, there's no function-like entity that I can return
(and I'd argue that's a bad idea anyway). In such cases, I've written
the error message so that it says '(none for N)'.