Emitter: EPIPHANY! Call piping should fake being a call. Fixes #145.
I've had a fair amount of problems doing call piping. Notably, the
issue of piping to a variant failing. So, instead, I'm going to have
func pipe trees pretend to be calls and then run a call. This
incidentally solves #145 which was caused by call piping not taking
variants into account. Now that func pipe shoots to variants, that's
no longer an issue.