It's finally over: Variants now build to enums. Fixes #148.
What's new:

Variants previously would get created as tuples, and then be lifted
into an enum of some kind. This was a little messy, because it meant
that there were two things around: One, the variant, and two, the enum
that was to receive the value.

I've made some adjustments in emitter (fewer reboxers, bigger gap
between partial and non-partial), ts (generic solving may unify if
there are ? types), debug + vm (enums are now instances, and marked,
collected, destroyed appropriately).

Why:

Option is an important enum, and will be dynaloaded into the core
during this release cycle. Part of being able to do that is making the
interpreter able to easily create enum values and to adjust their
contents as needed.

Suppose I have a function like Option::and_then. It needs to apply a
function to the Option but only if that Option has a Some inside.
That sounds easy but under the previous structure, there's two things
that I'm dealing with: One, the Option, and two, the Some. Since both
of those may need gc markers, I need to make sure that I have the
right type for both at all times.

Oh dear. Well...I know there's going to be a result register for
Option::and_then. That will have the new Option type that I need to
have made. But if I have a Some inside of the Option...oh no. That's
very bad. I need to CREATE A NEW TYPE AT RUNTIME. That is not a very
good thing to do at all.

But what if Some's didn't get made at runtime? What if it was just the
Option? Well, then the problem...wouldn't exist at all, now would it?
No.

That's why this is important. This single issue (which took a long
time), has been THE barrier to dynaloading Option and Result.