Empty variants get default values...except they're instances now.
This is useful because, for example, the None of an Option doesn't
need multiple instances lying around to define itself. Before variants
built out to enums, this was being done but it was a plain integer
literal instead. I've decided to make it an instance so that code can
depend on that instance-like entity being there.