Enums as optional arguments! Neat!
This includes a fix for a recent commit that added backing literals to
empty variants. That code was wrong and broke if there was a variant
that had concrete types. This includes a fix + test for that.

Anyway, it's now possible to have enums as optional arguments provided
that at least one variant is completely empty (such as the None of
Option). This is partly for simplicity (no exotic invocations of
variants as a default), and partly because it requires no changes to
the vm or debug to work.

One use for this is that you can effectively have default arguments of
any type: Simply make it an Option[whatever] and give it None as the
default. The caller has to do an added Some(x) instead of (x), but in
return you can have defaults. One example of it is how one test
creates a my_print function that allows specifying a file target (with
stdin being used if None is set).

I added this feature because I liked seeing what Python could do with
optional arguments and its print function.

Finally, I'd like to note that this commit is the reason why I added
backing literals to empty variants. I couldn't do this without them.