Generics update HAS to come before block entry.
In trying to do Option dynaload, I accidentally stumbled onto this
problem. What happens is that Option is the first case where something
dynaloads that is a class and has generics. So in certain cases, if
there are 2 generics but Option dynaloads, it'll want to set the
number of generics to 1.

The problem comes in that the block was entered BEFORE the update. So
when Option goes to restore generics, it restores to the caller's
caller's generics. So fix that by updating generics as stated above.

I have also removed the generics update from the bottom of enums. I
don't think it's necessary since function block exit will update the
generic count automatically.