Get rid of the last use of lily_ts_easy_resolve.
This was being done because a long time ago type_matchup was in charge
of doing any reboxing. Since it no longer does that, having two
versions of a wanted type is no longer necessary.

This turned out to be the last use of lily_ts_easy_resolve, so I've
decided to get rid of that too. In a prior commit, I mentioned that
lily_ts_easy_resolve is almost always wrong (it'll return NULL where
it really should default to any).