It's time to drop show.
'show' is one of Lily's earliest builtin functions. The premise of it
was rather simple: One function that can dump any kind of a value.
Given that Lily did not (and still does not) support a common protocol
for doing a .to_s or .inspect or .repr, this made sense. More
importantly, show was used as a means of dumping the bytecode of a
function in a way that made more sense.

Initially, that representation was pretty lame. But I eventually went
back and redid it. That ended up being how show looked for a year
plus. For a long time, show was great. I made sure it was always
updated with the vm and it provided useful analysis to determine if
something was a emit-time problem or a vm-time problem.

Times have changed.

* Bad emitting is unheard of now that there are so many models of how
  to do things right in emitter.

* show's biggest feature, the ability to display anything, has been
  rendered less useful now that string %s can do any type. More
  importantly, string interpolation and better formatting will be an
  even better alternative to show.

* "How do I show something that's a foreign value". Well, see, you
  don't. That's the problem: The proper way to do this is how (at
  least) Python and Ruby do it (have classes get an automatic to_s or
  let them define one, and use that).

* As far as most instances go, I really don't care what values are
  inside of them or what the private/protected/public setting on them
  is. Just dump the named values out (hi ruby).

* function opcode dumping is getting REALLY bad to keep up with. The
  new typeless_vm branch will finally cripple show by (likely)
  allowing for storages to change their register types, thus making
  show near impossible to work with. Furthermore, the new switch op
  in typeless_vm does not have type information, making show utterly
  unable to list what variant that each jump corresponds to.

I'm keeping some of the leftovers that show needed (variable names
packed in functions), because I want to fix all that after pulling in
typeless_vm.