Merge typeless_vm. This fixes #149.
This commit message will be very long, because this is a huge change.

From the very beginning, the vm has always had type information. A
value has, until now, consisted of a full type, flags, and a value. It
was great until generics came along.

When generics came along, so too did the idea of solving generics at
vm-time. I made that choice because I did not want to make specialized
copies of functions on-demand, either in emitter or vm. I chose
runtime solving to save memory, the same reason I've made many
decisions in Lily. An alternative would have been to generate unique
specializations for each invocation, but that seemed inappropriate
with Lily's handling of primitives (in C++, this makes more sense).

For a long time, that worked well enough. But then the problems
started showing themselves.

* Creating values requires a type which has to come from symtab, vm,
  or parser.

* Suppose a function closes over an upvalue and that upvalue is
  unsolved. The closure may need a register to have a type but have
  no access to that type. This is the and_then bug, and it was solved
  by type blocks.

* Making new registers requires a non-derefable type.

* Foreign calls may require solving, and need a resulting type.

* Runtime solving means building new types at vm-time. Lily's types
  are supposed to be unique, and this in turn means that such an
  operation is not thread-safe.

Removing types from the registers gives a lot of benefits though:

* Fewer registers per function, since they can be re-used without
  keeping their type fixed (but still keeping track OF the type).

* It's now easy to have an special assign that rewrites the type of
  a register.

* Generic functions now have zero solving cost.

* No more 'does this really need to be tagged' nonsense (sadly,
  because tagging is more aggressive now).

* In theory, this leaves the frontend to handle the typing while the
  vm handles the more complicated parts. In the future, this may be
  useful for allowing other frontends.

There are some downsides to erasure (notably, casting away from any is
much less useful). However, I feel the performance benefits strongly
outweigh the benefits in this situation.