New!: '<class>(...)' to construct new values.
It's still possible to use <class>::new, since I haven't decided if I
want to get rid of that or not. This was done so that the new Dynamic
class can be opt-in but without having numerous Dynamic::new calls.

So far, this works by implicitly targetting the ::new of a class. In
theory, if foreign classes export a ::new, then this should work on
them.

Enums, even if a ::new is incorrectly created on them, are not allowed
to use this feature. I consider it wrong for an enum to be created
through this kind of strategy, at least for now.