any is now Dynamic, casts are now safe. Fixes #150.
This is a huge change, because the new Dynamic class is opt-in instead
of being a type that is fallen into. Doing so allowed removing a lot
of special cases in emitter for defaulting to any. Defaulting to any
is...pretty much dead now. I'll probably fix it to default to a new
Nothing type in the future.

The new Dynamic type makes use of the previous commit allowing a class
name to be used as a constructor. I adjusted quite a few tests to add
uses of Dynamic where needed. A few were removed, since the new
Dynamic class allows nesting (I decided for this after seeing Haskell
do the same).

On a side note, casts only work against Dynamic now. They are useless
against other types, and have been ever since unification was baked
into ts. I took another page from Haskell and made casting yield an
Option of the given type. By doing so, BadTypecastError is now
completely gone.