Double is no longer hashable, and...
Hashability is now set when loading the builtin package. It no longer
comes from the flags, and so I removed that field from the class seed.
I decided to condense the size of that field down.

I decided to make Double non-hashable because C's double type is not
precise. NaN, in particular, is always non-equivalent to itself,
making a very, very bad hash key.