Build a ?-padded enum instead of a partial variant type.
The padded enum type isn't as useless as the partial variant, because
there's a slim chance the partial will get returned in a generic call.
The real point of this, however, is to make it where there is
(eventually) no reason to build types of a variant since bare variant
values are banned anyway.