Use the actual self type for method lookups.
The old line assumed that the top of the class chain was equivalent to
self. This is not true when parsing the methods of an enum. In that
case, it's likely to be a variant class. But the line worked anyway
because variants always have their enum as their parent.

The new line goes straight for the right class, and will not attempt
to find methods within a variant.