Pull variant classes away from regular ones.
lily_class has been treated as a garbage dump for far, far too long.
It holds info for builtin classes, user classes, foreign classes,
enums, and variants. What this commit does is to rip away the variant
class part away from regular classes.

There's a slight hack right now, in that variant classes are added to
the class chain of an import (they get withdrawn later for scoped
enums). The parser doesn't allow bare variants, so this is (as of now)
safe to do. I may, in the future, have a linked list of variants that
is stored away from the regular classes though.

In a test case of 'var v = Some(1)', this results in a (one-time)
saving of roughly 200 bytes.