Finally get postgres working again (many fixes).
A number of things have been bundled with this:

* Result.fetch_row and Conn.new are gone.

* Added Conn.open, Result.row_count, Result.each_row, and
  Result.close.

* Added new api for creating no-ref enums.

* Dynaloading had a bug where methods would be attached to the wrong
  area. So that's been fixed.

* When a library is loaded, the number of classes+enums it needs are
  set as reserved ids. Those ids are then filled in as needed.

For now, the postgre bridge uses Options. This is temporary: I want to
do some value api adjustments before fixing that. It'll eventually use
Either instead of Option.

This commit sucks: It's a huge fix because nothing exists to check the
stuff except postgres, and postgres needs a lot of stuff to work
again.

That old postgres.Error class is gone, btw.