New strategy and naming for Option+Either.
The old strategy was to ref something but shallow copy it into the
enum. In practice, that turned out to be wrong nearly 100% of the
time. In no case did I find out that it was correct after a closer
leak inspection.

Now, the strategy is to shallow copy, but not ref. In practice, this
works perfectly for just-made values. In cases where that doesn't
work, lily_copy_value can be used to clone the value over.

I also made Some, None, Left, and Right creation less verbose.