Lexer: Line readers should return the line length or 0.
Previously, they would return either 1 or 0 (success or failure).
However, having the line length would be nice in the future for string
and lambda handling. Both could, in theory, run a single size ensure
after scooping a whole line. It's then just a matter of ensuring that
pos + line_length is handle-able.

Doing the above would eliminate grow checks that fire off for each
character that's scanned in.