Lexer: Your entries don't need read/close funcs.
A long time ago I wanted Lily's line reading to be more generic and
pluggable. It was a stranger time. For a time, it even used apache's
reading functions to help scoop text from a file.

Eventually, I got bored of that. But around that time, I also gained
different kinds of string reading (lambdas need to copy text). Because
of that, the different kinds of entries stayed.

Well, no more. Each entry gets an enum marker, and that marker is used
to determine how the entry is read from and gets closed.