Lexer: Call ensure_label_size with a full line width.
The old way was to check for the label size being ok with each char
being scanned. No more! The previously-promised 'ensure for each line'
has now been implemented. There's a slightly ugly case where, for
comment blocks, that it uses strlen of the line because there's no
max line length information left. That's okay, because block comments
shouldn't be that frequent in a lambda (and the line that comes after
shouldn't be extra large). It's a worst case that I don't see being a
big problem, because most lambdas are small.

Most of the checks have a +3 in them. That's to make sure that lambda
scan has room to add '}' and '\0'. If it ever comes that such a thing
is not necessary, then +3 can be reduced or removed. It's likely
overly aggressive, but I'm okay with that right now.