Interpolation has landed. Fixes #143.
I initially considered punting this off to the next release. But then I
realized that printfmt and String.format both now take `List[Dynamic]` and
that Dynamic (unlike any) is explicit. Casting to Dynamic is really awful, so
it just had to get lost.

Interpolation is marked by `$" ... "`. `$""" ... """` works too. In either
case, the interpolation blocks are denoted by `^( ... )`, with parenth
matching done to ensure a balance. Interpolation blocks can contain strings,
and pretty much anything so long as it comes out to a single expression. I
intend to add more restrictions later (no lambdas?)

With this, it's now possible to get rid of printfmt and String.format.