Farewell, String.concat.
It's a bit sad to see it go, because String.concat was the first real
method that I put into the interpreter. It was initially written
because it was really easy and I needed to test class methods. But as
time has passed, this method has found itself to be less and less
important.

With the arrival of interpolation, it became simply impossible to
justify this staying around. Interpolation has the advantage of being
an opcode, so it doesn't incur the function call overhead of
String.concat. Moreover, interpolation is quickly becoming more
idiomatic in my mind (insofar as how I want Lily to be written).

Farewell, String.concat.