Add new string functions to lily_value.{c,h}
I was hoping to wait until the next release for this, but I can't.
There are too many paths that are creating raw lily_string_val's, and
each one is a potential problem.

This commit just tidies up what's in lily_value. The new functions are
not used outside of it yet. I've done this so I can later verify that
the newly made functions are not broken (it's their callers).