A String inside a container should be quoted+escaped.
When I fixed strings to not escape at all, it came out a little weird.
Specifically, a String held in some sort of container, like a List,
would not have the quotes printed, and would not be escaped. I looked
around, and both Python and Ruby at least do this. So it made sense to
do it too.

Add toplevel Strings directly, quote those that are nested.