Add tools/ and dynascan.py
dynascan.py can be used to pull dynaload information from a file with
the right kind of comments. It currently only understands define
blocks, but I suspect that it will be extended to more in the future.
I've built it to make that easier.

The tool is currently unused, but I'll soon be using it to start
generating the dynaloads for the builtin classes.