Import names now based on paths, put main import in bootup.
Two in one, because they both rely on each other a lot.

The first is simple: The loadname of an import is now based on the
path that is provided. This means no more passing a loadname and a
path: Only a path is needed now.

The second, is that more of the first import is creating during boot.
The parser loop now only fixes up the name on the first pass, and
doesn't have to do as much fixing up.