New! Import can now have '/' in paths!
This is huge for the language, because it makes having modules into
something realistic. The last few prior commits have been builting up
to this.

The syntax for import is the same, just that it's now possible to do
'import x/y/z'. For now, having a trailing slash is not allowed, but
that may be allowed in the future as a means of importing a whole
module. I'm not sure yet.

Paths provided are not quoted. This prevents using string scooping
which would implicitly allow '../'-like paths. I don't want that.

Do note that import paths are relative to the current import.