Revert "Don't "fix register types" except for optargs."
This reverts commit 4f70bbf95d3f54b5f70507b8861e3ab4cb64d649.

This is unfortunate. It is a temporary fix to a longer-term problem.
The problem is that I want very fast arithmatic operations that do not
need to check register types. I also do not want to fix register types
on each pass.

As it stands right now, I cannot have both. I have uncovered a bug
where the above-mentioned patch causes a leak when an arith operation
targets a storage that never got zapped.