New! Speculative gc. Less tagging, more speculating.
This touches most of the interpreter, and marks a huge change in Lily.

The centerpiece is a new tag, VAL_IS_GC_SPECULATIVE. If this is set on
a value, then that value is gc walked. However, values that are set
with VAL_IS_GC_SPECULATIVE are not given gc tags.

This includes some new machinery in parser to write down that a class
is either basic (no tag, no speculation), speculative (needs that new
tag), or tagged (must have a tag). Two new opcodes have been added so
that class construction can be specialized.

I've changed how build values are stored in stack. It holds the full
value now, so that I can assign the value over instead of doing a
fixed move.

The moves in src/lily_value.c have been changed so that they are all
speculative by default. That will be addressed in a future patch.

This removes a lot of tagging that was previously being done. That
tagging has been replaced by speculation (or in some cases by using
basic classes which avoid gc tags altogether).

Finally, this leaves a lot of low-hanging fruit. Static lists, hashes,
and enums are all created speculative. Moves are speculative when they
may not need to be. A value marked speculative is not unmarkable. In
the future, those things can be rectified for increased performance.