Unify collect + destroy functions under one roof.
Previously, gc collect functions and destroy functions were two
different beasts. These were, for the most part, copy+paste with a
little bit of change between each class. They weren't very error prone
(but I did notice that collect seemed to be forgetting to destroy
files properly).

This patch unifies collect + gc underneath one banner. It's probably
slower by the slightest of hairs due to having to do extra gc value
checks and a sentinel check at the top. This is still better.

I've also added a gc sentinel as a guard against re-entrancy. It
allows the gc to set last_pass to -1 to indicate the end. The first
time that is seen, the new destroy will hollow the value and swap out
to use the sentinel. When the sentinel is seen (checked at the top),
nothing is done at all.

The values of the sentinel are meaningless, but they can't be 0 or it
gets "optimized" as a null pointer which is wrong and bad.