Complex moves now have a _f suffix. MOVE_* flags.
With the new gc strategy recently introduced, there's now a need for
moves to be more particular. Are you moving something that is going to
be uninteresting to the gc? Be explicit about it!

Most moves now have a _f suffix if they take flags. I've added four
flags to lily_value.h to indicate what kind of value. I've simplified
the moves (no more shared_enum, instance_gc, or closure) as well.

The new move strategy makes it harder to be wrong, since the caller
sends only flags that show deref-ability/gc interest level, and not a
type indicator flag.