Add lily_new_empty_value, kill lily_new_value.
Previously, creating a new value meant specifying the VAL flag for it
and then casting a value to a union. Pretty ugly stuff. With moves now
more stable in the core, stop doing that.

The new way is to create a blank value and use the right move to push
whatever you've got inside. This new strategy has two benefits:

* It's harder to get wrong, since moves specify the type flag.

* There are less warnings now.