Put sipkey into options, create lily_api_options.h (toward #175)
This commit pulls lily_options out from parser and lily_core_types,
and into the new lily_api_options.h. For now, lily_api_options.{c,h}
contain just the struct, default create, and destroy. It may or may
not be expanded upon in the future.

This also moves the sipkey away from vm and into the option struct.
It's still the same dumb key as before. On the positive side, this
makes it more feasible for an outside caller to specify a non-dumb
sipkey.