Drop lily_impl.h. This file has never ever contained much.
When I first made Lily, I added it in because I expected that Lily
would end up needing a lot of outer functions provided to work. Time
has shown that to not be the case, so drop the file.

Lexer, now the only user of it, declares it at the top and explains
what it's for. I call that good.