Some basic rewinding, and a partly-working repl!
This introduces lily_load_chunk, which will rewind back when there is
an error. This loading is transparent to the caller, because I think
the caller is not in the best of positions to direct the interpreter
about how to rewind. It's still experimental.

The repl is half-working right now. It doesn't do continuing, and it
does not do vm rewinding yet. Those are yet to come.