NEW! Here's that packaging system, as promised. (Toward #178).
This is a complete rethinking of how Lily organizes code. The
importance of this cannot be understated.

Before this commit, extra code could come from a variety of hardcoded
directory paths. This was great for convenience: You compile some new
library for Lily and then drop it in one of the directories. Easy,
right?

What happens when two packages want different versions of a library?
Before this, that was never taken into consideration.

Now, Lily has a new keyword, called 'use'. 'use' loads packages from
a 'packages' directory relative to the first file loaded (or ./ if
running from -s). Built-in packages are now accessed through 'use'
instead of 'import', so there's bound to be a lot of naming changes.

With this change, now two different packages will have their own
package directory, and can have two different versions of the same
library if they so wish. Two big benefits of this are that it makes
packages modular, and it makes for more easily repeated builds later
down the line.

The package manager is yet to be done, but this does all of the
internal work to get it ready.