Fix strange type issue in try/rpn_calculator.lly
Dynaloaded enums were not registering the number of generics they had
available. ts NEEDS that information, because it sets aside space for
generics to solve into by NULL-ing out slots.

Since that information was not provided, an Integer type was bleeding
down, causing the Left to solve with the wrong type.