A small adjustment to package path lookups...
Previously, libraries were checked as if they were directly in the
packages directory. That's not consistent with how plain files are
loaded. Also, plain files had a specific name, and I didn't care for
that either.

The new paths are:

`packages/$NAME/$NAME.lly`
`packages/$NAME/liblily_$NAME.(so/dll)`