Make lily_free a function, not a macro.
It's hard to guess the future from where I'm at. But I could see
myself say, compiling Lily with a more aggressive malloc
implementation. If libraries are using the regular free instead of
Lily's new free implementation, then that would make them unable to
participate in testing of it.

Later on, it can be macro-ized later if I need or want to do that
again.