Move (most) destroy functions into lily_value_ops
This means that lily_cls_* files are not being included just to get at
the destroy functions inside of them. This strategy also makes it
easier to see a full reference implementation of destroy functions
(except hash) to make sure they're correct.

Hash is not included because a couple functions within hash require
being able to destroy values. I'll probably merge it in later, but if
I do, I'll do it after redoing hashes internally.