Take the membuf away from ast, and give it to emitter.
Emitter is now fully in charge of rewinding the buffer back, which it
does whenever it runs an expression. Currently, there's a small issue
wherein parser doesn't rewind the buffer when doing freeze/thaw of
states. I plan to fix that later down the line, when I can pull ast
state away from the pool.