Remove the repl.
The repl is missing a lot of functionality. It's also very common that
people are going to look for/ask for a repl. I don't want their first
experience with Lily's repl to be a broken mess. Which is what it is
right now.

I fully plan to add the repl back in. But for the time being, there is
a lot of internal refactoring that needs to be done. That refactoring
will alter what and how data is going to be saved/restored.

Sorry.