Create new o_get_integer opcode, and some partial use.
Before this, && / || were implemented by loading up a literal with a
value of 1 and a value of 0. That's not the best solution, as you can
imagine.

So now there's a new opcode: o_get_integer, that does that. I know
that conditions should be Boolean, but I'll get to that later. For the
moment, this opcode helps to wean emitter off of literal information.