Fix a (recently introduced) bug involving return and exceptions.
The bug was caused by moving up the try block pop when processing a
return statement. If the expression sent to return was capable of
throwing an exception, the exception would not be caught because of
the try block pop.

This fixes that, and also adds a test just to make sure I don't goof
this up again.