Replace o_setup_optargs with o_optarg_dispatch. HOWEVER...
First, the rationale:

Previously, optional arguments worked by having a counted set of pairs
of register spots and literal spots. For a long time, this was totally
okay in the vm. But...then I added variant values as optional
arguments. So it seemed natural to have a 'backing value' as a literal
so that optional arguments would work A-OK.

There are three problems with this approach:

* The new opcode o_get_integer, can be used for loading small integer
  literals...but not by optional arguments (since they require
  literals).

* Variants require a backing literal. This cascades down to foreign
  functions, which have to follow about 7 pointers down to get the
  backing value for, say, a None value.

* It's a very strange opcode, as it's the only one with a series of
  pairs inside of it. That makes it difficult to work with in terms of
  closure patching and some extra stuff I want to do later.

What the new opcode does:

It's pretty similar to how o_match_dispatch works. It takes a series
of jumps, and follows the appropriate one based on which argument is
the first one to have a value installed. Since it works based on
jumps, I can write whatever I want in between. Like new opcodes or a
better way of variant defaults.

The bad part:

Closures no longer work with optional arguments. Some inspection
reveals that this has actually been happening for match dispatch as
well.

For now, I've decided to comment out the offending test. I don't like
doing that, but this is a unique situation where I'd really prefer to
start destroying the closure code instead. It's really fragile.