'use' now checks libraries using `name` instead of `liblily_name`.
I really don't like the liblily_ being a special case apart from all
of the others. If something wants to eventually use the libraries as
actual libraries (ex: for linking), then it can add the lib prefix it
needs.

Also, these libraries shouldn't get mixed around with normal system
libraries, so I'm not really seeing a need for the special prefix. I
may revisit this in the future though.