BIG! Dynaload 2.0, cid tables, fixes #181.
This is a huge change to the language that does many things that have
been badly needed for some time.

Dynaload tables are now backed by strings, instead of a linked list.
This means their implementation is a lot saner, and that there are no
seeds to worry about the special bits of. Sure, there are custom
internal listings, but they're simpler and easier to spot now.

Runners or those registering a package now need to define a loader. A
loader is reponsible for transforming an integer id of some entry in
the dynaload table into either a function or a var. This is probably
slower (Lily used to grab the functions from the seed of the linked
list entry), but it's for the better. Libraries like postgres that
only specify classes and methods can have Lily automatically load
through a naming scheme of `lily_$PKGNAME_$CLSNAME_$FUNCNAME`. So
there's some savings there.

Then there's cid tables, cid being short for 'class_id'. These tables
contain integer ids of dynaloaded classes, in a format that the
dynaload table itself specifies. The usefulness of cid tables comes
from having the right ids at vm-time, regardless of if something was
dynaloaded or not. They aren't used much now, but will be in the near
future.

This also merges all `lily_cls_*` files into one lily_pkg_builtin file
so that Lily's loader doesn't need to do a boatload of externs. I've
also gone to the trouble of making Tainted's sanitize done through a
built-in function instead of being bootstrapped.

I've done some work adjusting postgres and apache for the new internal
stuff (cid tables and new dynaload tables). For now, their dynaload
tables are hand-crafted, but in the future I hope to have a tool to
automatically generate them. I've added some nasty stuff to apache in
particular, because I haven't gotten around to rewriting hash stuff
just yet.