ts: lily_ts_check should use covariant matching.
Using invariant matching (the old way) caused a container that
inherited another to not properly match. This seems to be the correct
thing to do overall, anyway.

Note that ts still doesn't pin a generic when a generic is found in an
invariant position. But I'll get to that later.