Introduce scoop types as ~1, ~2, plus Tuple.merge and Tuple.push.
The scoop types are a solution that allows any arity of Tuple to be
used with the two new Tuple functions. When a scoop is on the left, it
will greedily match as many types as it can on the right side. The
important part is that there's no matching against some bottom type,
so there's no degrading of the types that come in. Also, the types are
saved.

Resolving then has all of the types that were sent in. This means that
`Tuple.merge` works for any combination of two Tuples, and still
retains their outgoing type.