Create lily_vm_raise, which handles vm dynaloading.
Previously, the vm's exception handling code did the 'do I dynaload
here?' logic. This was very, very bad because that's a hot area that
should not have to deal with that concern at all, ever.

Parts of the vm which previously called lily_raise now instead call
lily_vm_raise which makes sure that the exception matching the given
id has some class associated with it. If it doesn't, the dynaload is
run.

The downside to this is that the prep for __main__ way down at the
bottom of emitter has been doubled to 32. This was done because I'm
not yet ready to tackle the underlying problem (I need some other
framework first).