lily_buffers -> lily_buffer_16.
Internally, this refines the idea previously had of a generic buffer
type that can handle various types. It was a pretty weaksauce
implementation of templates. Now, it's just a buffer of uint16_t
(shortened to u16). The focus comes along with an added writing
function for writing two at once.

The new writing function helps to eliminate some minor issues that
were present before (optargs liked writing pairs, now it can). But
this also lays some foundation for making emitter's code into a
code buffer, instead of it being spread out into the emitter struct as
it is right now.