Emitter's code is now a lily_buffer_u16.
Writing functions that were in emitter (1 to 6) are now within
lily_buffer_u16, and thus available to the other u16 buffers.

A key motivation for this is trying to clean up how the buffers are
used internally by emitter. This makes it easier for, when needed, to
have parser do a smidgen of code writing. It also discourages direct
writing, which has been the source of a large number of problems.

This is also a step toward rewriting closure patching.