msgbuf: Clean up prototypes, make header not need lily_core_types.h
Mind you, the actual implementation of msgbuf needs lily_core_types.h still.
But the outside no longer needs that, which is important because msgbuf is
used by the frontend as well as the backend of the parser.

Changes including dropping lily_msgbuf_add_type (use add_fmt/add_fmt_va
instead, as you're more likely to actually need that), and making passing a
bytestring to passing the buffer and size directly.