Create o_foreign_call and o_native_call.
There are now three opcodes for calling a function. The difference
between them is the source of the call:

o_native_call:
    The target is the id of a 'foreign' function, such as print.

o_foreign_call:
    The target is the id of a function created through `define`.

o_function_call:
    The target is a register, which may hold either of the above.

Lily does not allow built-in functions or functions created through
`define` to be assigned over, so detecting this is very easy to do.

Having this distinction allows each call opcode to focus on exactly
the case it has, so there's no longer a need for o_function_call to
have a 'call style' field in the bytecode.