Create o_except_catch, o_except_ignore.
This specializes the except clause code so that there's no longer a
need for a field that just says if the code is to be caught or not. I
have also done some tuning to the actual exception catching around
the opcodes, since there was some silliness (setting the next jump
before checking success, for example).

Also, I fixed o_push_try, which falsely mentioned that it took three
values but then also said it had a count of two. The truth is it has
just two fields of interest.