Rewrite the front-page readme, which incidentally fixes #180.
Now, the front page has a postcard example, rather than just a bunch
of code. The sandbox is highlighted, and most importantly there's more
of a focus on why Lily should be picked over languages. It's a step
toward being more serious about the language...even if the examples
are all very inclined towards cats.

Unfortunately absent are the ~1 and ~2 scooping types (as well as
Tuple merging). I'll make a note of that at some point in the future.